project1-chatbot/requirements.txt
# Core LLM and NLP libraries
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# LangChain for conversation management
langchain>=0.1.0
langchain-community>=0.0.10

# UI Framework
streamlit>=1.25.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Vector store for context retrieval
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
